requirements.txt
requests
python-dotenv
instaloader
